Add mockup icon image.
I'm thinking something like this icon will work well for
HideSettingsPages. I made it in mspaint and will have to make it look
better before using it as the app icon. The grey sidebar is supposed to
represent the Windows Settings app when it's open beyond a category and
the black lines on the right are for text. The red "X" at the
bottom-left is to show pages that have been hidden.