Replace Unix line endings with Windows...
line endings for the hsp_registry-helper info text file resource. This
makes it so that the text isn't all a mess in the About window if the
user builds from the master branch (or any other branch) hosted on
GitHub. It should display correctly with Visual Studio without doing
this, but it's better for the text to display correctly if someone wants
to test the latest code or build it from source when releasing the app.