Add new project to help with testing page lists.
CrLf2Semicolon will allow the user to paste in a page list on the left,
click a button, then have all the CrLf characters replaced with
semicolons on the right. This makes it easier to see which pages are
missing in the HideSettingsPages page list when it's applied to the
Registry key value manually.

CrLf2Semicolon will also have a checkbox to replace Lf characters as
well as CrLf characters. There will be "Select all" buttons for both the
input textbox and the output textbox.

For now, there probably won't be as much UI polish for this app as for
HideSettingsPages and hsp_registry-helper, but only because this app
isn't really meant to be used as much. If something bugs me with the UI,
I might fix it.