Allow copying value into directly apply window to work...
if "Shift" is held while opening the window or when the option is enabled.

See also #52.

This needs to have an enable/disable thing for the Options menu as well as a tooltip. Also, there needs to be a piece of code to do that when "Ctrl+Shift+D" is pressed.