Use an actual SaveFileDialog for exporting selections.
For now, I guess this makes it so that this can Closes #1, finally.

Might be a good idea to allow importing the current value from the Registry, so that the user can change what's already been written, if they want to.

Allowing exporting the file to XML may also be a good idea, but that might not be implemented for now since a text file seems to work fine as-is. Kinda like some other things that are fine as-is but might be better if they were the other option, though this would be an additional option in XML exporting and not removing support for using text files, unlike the other thing that kinda changes a lot of things.