Change name for About window banner and About...
tab's table layout panel to match the name of the same control in UXL Launcher. This allows for that particular tablelayoutpanel's color to be the same as if a theme were applied to UXL Launcher and the theme wants to use the tabpage color for that single tablelayoutpanel instead.

Having the banner get a different name allows the application's banner to show up instead of the UXL Launcher banner.