Increase the size of the Registry key value preview textbox.
Other changes in this commit include the Settings page list label being moved up slightly and aligning with the Settings page visibility label, as well as the Hide/Show Only radiobuttons being moved closer to their label. The buttons at the bottom of the window have also been moved slightly as the container they're in was decreased in size to allow a larger Registry key value preview textbox.